+++ core/src/platform/OpenGL/OpenGLRenderer.h
@@ -10,7 +10,7 @@ namespace Renderer {
     void clear() const override;
     void clearColor(uint32_t r, uint32_t g, uint32_t b, uint32_t a) const override;
     void clearColor(float r, float g, float b, float a) const override;
-    void setViewport(int32_t x, int32_t y, int32_t width, int32_t height) const override;
+    void setViewport(int32_t x, int32_t y, float width, float height) const override;
 
     void drawIndexed(
       const std::shared_ptr<VertexArray>& vertexArray,
