+++ core/src/renderer/Renderer.h
@@ -11,7 +11,7 @@ namespace Renderer {
     virtual void clear() const = 0;
     virtual void clearColor(uint32_t r, uint32_t g, uint32_t b, uint32_t a) const = 0;
     virtual void clearColor(float r, float g, float b, float a) const = 0;
-    virtual void setViewport(int32_t x, int32_t y, int32_t width, int32_t height) const = 0;
+    virtual void setViewport(int32_t x, int32_t y, float width, float height) const = 0;
     virtual void drawIndexed(const std::shared_ptr<VertexArray>& vertexArray, const std::shared_ptr<Shader>& shader) const = 0;
     virtual void drawCount(
       const std::shared_ptr<VertexArray>& vertexArray,
